- created a Boss enemy that shoots bullets - Created a SmartEnemy - went through and converted everything from int to float in order to get the SmartEnemy to follow the player - Updated levels - Changed the color of bullets - fixed the sticky keys bug

+++ src/com/Jakob/BasicEnemy.java
@@ -18,7 +18,7 @@ public class BasicEnemy extends GameObject {
 
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int)x, (int)y, 16, 16);
     }
 
     public void tick() {
@@ -31,11 +31,11 @@ public class BasicEnemy extends GameObject {
         if (x <= 0 || x >= Game.WIDTH - 32) // reverse direction
             vX *= -1;
 
-        new Trail(x,y,ID.Trail,Color.red,16,16,0.03f,handler);
+        new Trail((int)x,(int)y,ID.Trail,Color.red,16,16,0.03f,handler);
     }
 
     public void render(Graphics graphics) {
         graphics.setColor(Color.red);
-        graphics.fillRect(x, y, 16, 16);
+        graphics.fillRect((int)x, (int)y, 16, 16);
     }
 }
